refactor(threads): split ThreadsService into smaller services
Created the following services:
- CommentService
- VotingService
- BookmarkService
- ThreadService
This improves modularity and separation of concerns.
