+++ src/threads/__http_tests__/threads.comments.create.http
@@ -1,9 +1,9 @@
 @port = {{$dotenv API_PORT}}
 @baseUrl = http://localhost:{{port}}
 
-POST {{baseUrl}}/threads/2/comments
+POST {{baseUrl}}/threads/1/comments
 Content-Type: application/json
-Authorization: Bearer 
+Authorization: Bearer eyJhbGciOiJQUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsInRva2VuVHlwZSI6ImFjY2VzcyIsInJvbGUiOiJTVVBFUl9BRE1JTiIsImlhdCI6MTc0NjQxMzYzMCwiZXhwIjoxNzQ2NDE0NTMwfQ.l8ZIt77xQPziSz-y9JGEpipPk6rhHDzfbMbTPyq-f_wB3mXD0DttHLUUe9-d0BmrUs76D6EPytPZ8llqbTsVNfKKYd83Xwx3dbQcVRnTcFO5o0d1wHqajytpjR-q6HVQAz_QzEcnpS8mFF22QkYoZdaFgFqjzNwGCpQmkr9bOc6OFGvd1MwSK1vgHYiwsUKYp0bmIVxOhU9pTmNpAzGCZODVgVC8T6J_zjZhSCrHASxMqMk7emOs7fW34dM-3PW4UCJ6LbjZ9X1fphoI0Q7wU4Usz7Vjw4SW9LNzr0qHLnCqK8USJM9Mdqmm3ER-S7xQoHqOCzC8FDCddZsudoErUUCnCdFEENbojwXXHXvmT5lSHIemDj2ki33xpk4XjfzD0IkLVqdz8UNrnnUB7trt6gwQa7O4XnMRQIQVyN8mh0tJxfWpelvdkIipaID2WIXP4f1q_kd6IiuFSOZOSOKKZyeiJLH3vI_6KJyh26cM65eQVhwpSBUn5D2AYqKt0BJPuPtn5HptoALOOBKYvuN3C2xZcHpSad78zafROlgjgt6fOltAhDjhWiE-pGO_mF4-nxnqP7XaDYxaxhYe1CAckgIZKlSkuYlhD8CjUWL9mBuT5f2nUiqO7X67XQWfHkoOKm2bYFn5eInQb-xVxAV09qevIJakfYVmMF2nNdjACNs
 
 {
   "content": "This is another comment on the thread"
