+++ src/threads/threads.module.ts
@@ -1,10 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ThreadsController } from './threads.controller';
-import { ThreadsService } from './threads.service';
+import { ThreadService } from './services/thread.service';
+import { CommentService } from './services/comment.service';
+import { BookmarkService } from './services/bookmark.service';
+import { VotingService } from './services/voting.service';
 
 @Module({
   controllers: [ThreadsController],
-  providers: [ThreadsService],
-  exports: [ThreadsService]
+  providers: [ThreadService, CommentService, BookmarkService, VotingService],
+  exports: [ThreadService]
 })
 export class ThreadsModule {}
